feapder/requirements.txt
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
better-exceptions==0.3.3
bitarray==2.8.2
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colored==1.4.4
coloredlogs==15.0.1
cryptography==41.0.4
cssselect==1.2.0
DBUtils==3.0.3
ddddocr==1.4.8
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.4.2
exceptiongroup==1.1.3
executing==2.0.0
feapder==1.8.9
Flask==3.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
humanfriendly==10.0
idna==3.4
influxdb==5.3.1
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
line-bot-sdk==3.5.0
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.0
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
outcome==1.2.0
packaging==23.2
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
playwright==1.38.0
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pyee==9.0.4
PyExecJS==1.5.1
Pygments==2.16.1
pymongo==4.5.0
PyMySQL==1.1.0
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==3.5.3
redis-py-cluster==2.1.3
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
terminal-layout==3.0.0
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.6
w3lib==2.1.2
wcwidth==0.2.8
webdriver-manager==4.0.1
Werkzeug==3.0.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
